Refactor getAllRequests method
