+++ server/routes/RequestRouter.js
@@ -1,9 +1,9 @@
 import express from "express";
-import { createRequest, getOneRequest, getRequestsWithin3Months } from "../controllers/RequestController.js";
+import { createRequest, getOneRequest, getAllRequests } from "../controllers/RequestController.js";
 
 const requestRouter = express.Router();
 
-requestRouter.get("/", getRequestsWithin3Months);
+requestRouter.get("/getAll/:user_id", getAllRequests);
 
 requestRouter.get("/:request_id", getOneRequest);
